+++ fabric/src/main/java/com/devnemo/nemos/vertical/slabs/DataGenerator.java
@@ -1,9 +1,6 @@
 package com.devnemo.nemos.vertical.slabs;
 
-import com.devnemo.nemos.vertical.slabs.datagen.BlockTagProvider;
-import com.devnemo.nemos.vertical.slabs.datagen.LootTableProvider;
-import com.devnemo.nemos.vertical.slabs.datagen.ModelProvider;
-import com.devnemo.nemos.vertical.slabs.datagen.NemosVerticalRecipeProvider;
+import com.devnemo.nemos.vertical.slabs.datagen.*;
 import com.devnemo.nemos.vertical.slabs.datagen.lang.EnglishLanguageProvider;
 import net.fabricmc.fabric.api.datagen.v1.DataGeneratorEntrypoint;
 import net.fabricmc.fabric.api.datagen.v1.FabricDataGenerator;
@@ -17,6 +14,7 @@ public class DataGenerator implements DataGeneratorEntrypoint {
         pack.addProvider(ModelProvider::new);
         pack.addProvider(EnglishLanguageProvider::new);
         pack.addProvider(BlockTagProvider::new);
+        pack.addProvider(ItemTagProvider::new);
         pack.addProvider(LootTableProvider::new);
         pack.addProvider(NemosVerticalRecipeProvider::new);
     }